fix: solved error and folder auth
