+++ src/router/MainRoutes.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import LoginScreen from '../pages/modules/auth/login/LoginScreen';
+import LoginScreen from '../pages/auth/login/LoginScreen';
 
 const MainRoutes: React.FC = () => {
   return (
